JA: Added tracked files

+++ src/main/java/com/qa/ims/controller/ItemController.java
@@ -39,7 +39,7 @@ public class ItemController implements CrudController<Item>{
 		LOGGER.info("Please enter the item name");
 		String itemName = utils.getString();
 		LOGGER.info("Please enter a price for the item");
-		Long itemPrice = utils.getLong();
+		Double itemPrice = utils.getDouble();
 		Item item = itemDAO.create(new Item(itemName, itemPrice));
 		LOGGER.info("Item created");
 		return item;
@@ -52,7 +52,7 @@ public class ItemController implements CrudController<Item>{
 		LOGGER.info("Please enter the item name");
 		String itemName = utils.getString();
 		LOGGER.info("Please enter a price for the item");
-		Long itemPrice = utils.getLong();
+		Double itemPrice = utils.getDouble();
 		Item item = itemDAO.update(new Item(id, itemName, itemPrice));
 		LOGGER.info("Item Updated");
 		return item;
